global refactor MainPage
